+++ features/merchant/power_merchant_subscribe/src/main/java/com/tokopedia/power_merchant/subscribe/view/viewcomponent/MembershipPmProCheckListView.kt
@@ -88,9 +88,9 @@ class MembershipPmProCheckListView : LinearLayout {
     }
 
     data class Data(
-        val orderThreshold: Int,
+        val orderThreshold: Long,
         val netIncomeThreshold: Long,
-        val totalOrder: Int,
+        val totalOrder: Long,
         val netIncome: Long
     ) {
         fun isEligibleOrder() = totalOrder >= orderThreshold
